Wire up organizations with security.
Organizations which are password protected can only be queried when the user password which was used for creation or the owner password (automagically generated during creatiuon) are present in the request header via the key 'Plooral-Credentials' in plain text form.

Protected organizations which are but not hidden will return a HTTP 403 when queried without a correct credential set. However, hidden organizations will return a HTTP 404 to avoid scanning for hidden organizations.